Prefer special-mode over view-mode
I'm bloody annoyed over view-mode's quit only quitting that mode instead
of staying in it and switching out the buffer or deleting its window (it
looks like its exit action does the right thing though).  As I only need
to ensure the buffer is read-only and has basic navigation bindings,
special-mode it is!